Refactor, move files to appropriate directories
