+++ bofire/data_models/strategies/api.py
@@ -1,9 +1,9 @@
 from typing import Union
 
 from bofire.data_models.strategies.doe import DoEStrategy
-from bofire.data_models.strategies.enting import EntingStrategy
 from bofire.data_models.strategies.factorial import FactorialStrategy
 from bofire.data_models.strategies.predictives.botorch import BotorchStrategy
+from bofire.data_models.strategies.predictives.enting import EntingStrategy
 from bofire.data_models.strategies.predictives.multiobjective import (
     MultiobjectiveStrategy,
 )
